requirements.txt
flask
lumibot
alpaca-trade-api
transformers
torch
python-dotenv
asyncio